MapLayerKey > moved keys to external config
